Add SYSINFO log level. Consolidate LoggerAdapter setup into a common function. Split log related objects into new aodncore.pipeline.log module.
