+++ aodncore/pipeline/configlib.py
@@ -7,7 +7,9 @@ from six import iteritems
 from six.moves import configparser
 
 from .exceptions import InvalidConfigError
-from .schema import LOGGING_CONFIG_SCHEMA, PIPELINE_CONFIG_SCHEMA
+from .log import (get_base_worker_logging_config, get_watchservice_logging_config, get_logging_config_for_watch,
+                  validate_logging_config)
+from .schema import PIPELINE_CONFIG_SCHEMA
 from .watch import get_task_name, CeleryConfig, CeleryContext
 from ..util import discover_entry_points, format_exception, merge_dicts, str_to_list, validate_type
 
@@ -174,93 +176,6 @@ class LazyConfigManager(object):
         self._watch_directory_map = None
 
 
-def get_base_worker_logging_config(pipeline_config):
-    """Get the *base* logging config for pipeline worker processes, suitable for use by logging.config.dictConfig
-
-    :param pipeline_config: LazyConfigManager.pipeline_config dict
-    :return: dict containing base worker logging config
-    """
-    base_logging_config = {
-        'version': 1,
-        'formatters': {
-            'pipeline_formatter': {
-                'format': pipeline_config['logging']['pipeline_format']
-            }
-        },
-        'filters': {},
-        'handlers': {},
-        'loggers': {}
-    }
-
-    # decrease log level for noisy library loggers, unless explicitly increased for debugging
-    for lib in ('botocore', 'paramiko', 's3transfer', 'transitions'):
-        base_logging_config['loggers'][lib] = {'level': pipeline_config['logging'].get('liblevel', 'WARN')}
-
-    return base_logging_config
-
-
-def get_watchservice_logging_config(pipeline_config):
-    """Generate logging configuration for the 'watchservice' service, suitable for use by logging.config.dictConfig
-
-    :param pipeline_config: LazyConfigManager.pipeline_config dict
-    :return: rendered watchservice logging config
-    """
-    watchservice_logging_config = {
-        'version': 1,
-        'formatters': {
-            'watchservice_formatter': {
-                'format': pipeline_config['logging']['watchservice_format']
-            }
-        },
-        'filters': {},
-        'handlers': {
-            'watchservice_handler': {
-                'level': pipeline_config['logging']['level'],
-                'class': 'logging.StreamHandler',
-                'formatter': 'watchservice_formatter'
-            }
-        },
-        'loggers': {
-            'watchservice': {
-                'handlers': ['watchservice_handler'],
-                'level': pipeline_config['logging']['level'],
-                'propagate': False
-            }
-        }
-    }
-    return watchservice_logging_config
-
-
-def get_logging_config_for_watch(task_name, log_root, level='INFO'):
-    """Get logging configuration for a single pipeline watch, intended to be merged onto the output of
-        `get_base_worker_logging_config`
-
-    :param task_name: name of the pipeline for which the config is being generated
-    :param log_root: logging root directory
-    :param level: logging level to
-    :return: dict containing handlers/loggers for a single watch
-    """
-    handler_name = "{name}_handler".format(name=task_name)
-    watch_logging_config = {
-        'handlers': {
-            handler_name: {
-                'level': level,
-                'class': 'logging.FileHandler',
-                'formatter': 'pipeline_formatter',
-                'filename': os.path.join(log_root, 'process', "{task_name}.log".format(task_name=task_name))
-            }
-        },
-        'loggers': {
-            task_name: {
-                'handlers': [handler_name],
-                'level': level,
-                'propagate': False
-            }
-        }
-    }
-    return watch_logging_config
-
-
 def load_pipeline_config(default_config_file=DEFAULT_WATCH_CONFIG, envvar=DEFAULT_CONFIG_ENVVAR):
     config_file = os.environ.get(envvar, default_config_file)
     config = load_json_file(config_file, envvar=envvar)
@@ -302,10 +217,6 @@ def load_json_file(default_config_file, envvar=None):
         raise InvalidConfigError(format_exception(e))
 
     return config
-
-
-def validate_logging_config(logging_config):
-    jsonschema.validate(logging_config, LOGGING_CONFIG_SCHEMA)
 
 
 def validate_pipeline_config(pipeline_config):
