+++ aodncore/testlib/conf/pipeline.conf
@@ -8,7 +8,7 @@
     "wip_dir": "/tmp/probably/doesnt/exist/wip"
   },
   "logging": {
-    "level": "INFO",
+    "level": "SYSINFO",
     "pipeline_format": "%(asctime)s %(levelname)s %(name)s[%(celery_task_id)s] %(message)s",
     "log_root": "",
     "watchservice_format": "%(asctime)s %(levelname)s [%(name)s] %(message)s"
